app/requirements.txt
flask_cors
flasgger
RPi.GPIO
